Clean up run target.
