+++ crates/workspace/src/jobs/mod.rs
@@ -1,5 +1,4 @@
 pub mod install_node_deps;
-pub mod run_node_target;
-pub mod run_shell_target;
+pub mod run_target;
 pub mod setup_toolchain;
 pub mod sync_project_config_deps;
